Move isSafeUrl to utils/urlSafety to break circular dependency
