+++ packages/core/src/extension.ts
@@ -18,7 +18,8 @@ import { SourcesTreeProvider } from './views/sourcesTreeProvider';
 import { HistoryTreeProvider } from './views/historyTreeProvider';
 import { WatchesTreeProvider } from './views/watchesTreeProvider';
 import { WatchesStatusBar } from './views/watchesStatusBar';
-import { registerCommands, isSafeUrl } from './commands/commands';
+import { registerCommands } from './commands/commands';
+import { isSafeUrl } from './utils/urlSafety';
 import { ViewRevealer } from './services/viewRevealer';
 import { initLogger, setLogLevel, logger, resolveLogLevel } from './services/logger';
 import { getInboxUnseenCount } from './services/inboxBadge';
