+++ packages/core/src/test/commands.test.ts
@@ -1,7 +1,8 @@
 import { describe, it, expect, beforeEach, vi, type Mock } from 'vitest';
 import * as vscode from 'vscode';
 import { WorkItemState, type WorkItem } from '../models/workItem';
-import { registerCommands, isSafeUrl } from '../commands/commands';
+import { registerCommands } from '../commands/commands';
+import { isSafeUrl } from '../utils/urlSafety';
 import type { WorkGraph } from '../services/workGraph';
 import type { ActionRegistry } from '../services/actionRegistry';
 import type { ProviderRegistry } from '../services/providerRegistry';
