+++ packages/core/src/utils/urlSafety.ts
@@ -1,0 +1,9 @@
+/** Returns the parsed URL if it uses an allowed web scheme (http or https), or null otherwise. */
+export function isSafeUrl(url: string): URL | null {
+  try {
+    const parsed = new URL(url);
+    return (parsed.protocol === 'http:' || parsed.protocol === 'https:') ? parsed : null;
+  } catch {
+    return null;
+  }
+}
