+++ packages/core/src/views/watchesStatusBar.ts
@@ -1,6 +1,6 @@
 import * as vscode from 'vscode';
 import { WatcherService, WatchedRun } from '../services/watcherService';
-import { isSafeUrl } from '../commands/commands';
+import { isSafeUrl } from '../utils/urlSafety';
 
 /**
  * Status bar item that shows the count of active/failed watches.
